dependency: add libxml dependency
Starting in 2013, gem-ified standard library gems were only made available
through Bundler if the gem was specified in the Gemfile, see:
https://github.com/sferik/multi_xml/issues/42

As a result we started to get the error: "No XML parser detected. If you're using Rubinius and Bundler, try adding an XML parser to your Gemfile (e.g. libxml-ruby, nokogiri, or rubysl-rexml). For more information, see https://github.com/sferik/multi_xml/issues/42."

To fix this issue we need to specify an xml parser in our Gemfile. We're not going to be doing that much work within an auth Gem, so performance of the xml parser doesn't matter that much. But we picked libxml for performance anyways. (https://medium.com/neuronio/generating-xml-in-ruby-builder-vs-nokogiri-vs-libxml-vs-ox-757ce2d8f83c)